Replaced FiguresPainterLibrary console app project with a classes library project
