+++ FiguresPainter/FiguresPainter.csproj
@@ -82,7 +82,7 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\FiguresPainterLibrary\FiguresPainterLibrary.csproj">
-      <Project>{1562dd4f-48f8-4f30-b5ad-01e6519cf91c}</Project>
+      <Project>{14be3eb8-fb65-4e1e-be62-84c9ef23de6f}</Project>
       <Name>FiguresPainterLibrary</Name>
     </ProjectReference>
   </ItemGroup>
